refactor: move MocketSocket from mocket.mocket to mocket.socket
